CI: Move all unit tests to test/unit
Change-type: patch
